requirements.txt
cx-Oracle==6.0b1
naviance-client==0.0.1
python-env==1.0.0